chore: fix circular dependencies in the build task graph
