+++ npm/react/package.json
@@ -20,7 +20,6 @@
     "@types/semver": "7.5.0",
     "@vitejs/plugin-react": "4.0.0",
     "axios": "0.21.2",
-    "cypress": "0.0.0-development",
     "prop-types": "15.7.2",
     "react": "16.8.6",
     "react-dom": "16.8.6",
@@ -99,7 +98,10 @@
           "{projectRoot}/dist"
         ]
       }
-    }
+    },
+    "implicitDependencies": [
+      "!cypress"
+    ]
   },
   "standard": {
     "globals": [
