+++ npm/react18/package.json
@@ -17,7 +17,6 @@
     "@rollup/plugin-node-resolve": "^11.1.1",
     "@types/react": "^16",
     "@types/react-dom": "^16",
-    "cypress": "0.0.0-development",
     "react": "^16",
     "react-dom": "^16",
     "rollup": "3.7.3",
@@ -69,6 +68,9 @@
           "{projectRoot}/dist"
         ]
       }
-    }
+    },
+    "implicitDependencies": [
+      "!cypress"
+    ]
   }
 }
