+++ npm/vue/package.json
@@ -20,7 +20,6 @@
     "@vue/compiler-sfc": "3.2.47",
     "@vue/test-utils": "2.3.2",
     "axios": "0.21.2",
-    "cypress": "0.0.0-development",
     "debug": "^4.3.4",
     "globby": "^11.0.1",
     "tailwindcss": "1.1.4",
@@ -91,6 +90,9 @@
           "{projectRoot}/dist"
         ]
       }
-    }
+    },
+    "implicitDependencies": [
+      "!cypress"
+    ]
   }
 }
