+++ npm/vue2/package.json
@@ -60,6 +60,9 @@
           "{projectRoot}/dist"
         ]
       }
-    }
+    },
+    "implicitDependencies": [
+      "!cypress"
+    ]
   }
 }
