+++ npm/webpack-preprocessor/package.json
@@ -34,7 +34,6 @@
     "chalk": "3.0.0",
     "chokidar-cli": "2.1.0",
     "common-tags": "^1.8.2",
-    "cypress": "0.0.0-development",
     "dependency-check": "2.9.1",
     "deps-ok": "1.4.1",
     "find-webpack": "1.5.0",
